AI_Resume_Copilot/requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
ollama